Move context handling into events worker
